gateway-mt:  fix uint32 comparison on 32-bit architectures
The private Jenkins build is broken because of this issue.

miniogw/multipart.go:108:26: constant 4294967295 overflows int
failed to build: failed component "cmd/gateway-mt":
failed osarch {"linux" "arm"}:
build failed: exit status 2

+++ miniogw/multipart.go
@@ -105,7 +105,7 @@ func (gateway *gateway) PutObjectPart(ctx context.Context, bucket, object, uploa
 	defer mon.Task()(&ctx)(&err)
 	defer gateway.log(ctx, err)
 
-	if partID < 1 || partID > math.MaxUint32 {
+	if partID < 1 || int64(partID) > math.MaxUint32 {
 		return minio.PartInfo{}, minio.InvalidArgument{
 			Bucket: bucket,
 			Object: object,
